requirements.txt
langchain==0.2.14
langchain-groq
langchain-pinecone
pinecone-client
langchain-community==0.2.12
langchain-ollama==0.1.1
ollama==0.3.1
flask
pypdf
python-dotenv
pinecone-client==5.0.1
-e .